Separated ViewController and ViewModel in ReviewsViewController and ReviewsDetailViewController files.
