+++ TheMovieDb/ViewModel/ReviewsDetailViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ReviewsDetailViewModel.swift
+//  TheMovieDb
+//
+//  Created by Karla Rubiano on 12/11/21.
+//
+
+import Foundation
+
+final class ReviewsDetailViewModel {
+    
+    var review: ReviewsDetails?
+}
